// client/dist/widgets/common/print/src/config.ts
export interface ActiveItem {
  id: string
  title?: string
}

export interface OutputDataSourceWarningOption {
  label: string
  widgets: string
}

export enum PrintExtentType {
  CurrentMapExtent = 'CURRENT MAP EXTENT',
  CurrentMapScale = 'CURRENT MAP SCALE',
  SetMapScale = 'SET MAP SCALE'
}

export enum MapFrameUnit {
  Point = 'POINT',
  Inch = 'INCH',
  Centimeter = 'CENTIMETER',
  Millimeter = 'MILLIMETER'
}

export interface WebMapFrameSize {
  width: number
  height: number
}

interface CustomTextElementEnable {
  [key: string]: boolean
}
// ...
export enum ModeType {
  Classic = 'CLASSIC',
  Compact = 'COMPACT'
}

export enum PrintServiceType {
  OrganizationService = 'ORGANIZATION SERVICE',
  Customize = 'CUSTOMIZE'
}

export enum PrintTemplateType {
  OrganizationTemplate = 'ORGANIZATION TEMPLATE',
  Customize = 'CUSTOMIZE'
}
// ...
export type FormatType = 'pdf' | 'png32' | 'png8' | 'jpg' | 'gif' | 'eps' | 'svg' | 'svgz' | 'tiff' | 'aix'
// ...
export interface LayoutInfo {
  layoutTemplate: LayoutType
  pageSize?: number[]
  mapFrameSize?: number[]
  mapFrameUnit?: MapFrameUnit
  hasAuthorText?: boolean
  hasCopyrightText?: boolean
  hasLegend?: boolean
  hasTitleText?: boolean
  enableAuthor?: boolean
  enableCopyright?: boolean
  enableLegend?: boolean
  enableScalebarUnit?: boolean
  enableCustomTextElements?: boolean
  customTextElementEnableList?: CustomTextElementEnable[]
  enableNorthArrow?: boolean
  layout: LayoutType
  layoutOptions: {
    hasAuthorText?: boolean
    hasCopyrightText?: boolean
    hasLegend?: boolean
    hasTitleText?: boolean
    customTextElements?: any[]
  }
}

export interface PreviewOutLine {
  color?: string
  size?: string | number
}

export enum LayoutTypes {
  ServiceLayout = 'Service Layout',
  CustomLayout = 'Custom Layout'
}

export enum ReportTemplateTypes {
  RPTT = 'RPTT',
  RPTX = 'RPTX'
}

export enum ReportTypes {
  ServiceReport = 'Service Report',
  CustomReport = 'Custom Report'
}

export type PrintParameters = __esri.PrintParameters
export type MapView = __esri.MapView

export interface PrintTemplateProperties extends __esri.PrintTemplateProperties {
  templateId?: string
  printExtentType?: PrintExtentType
  wkid?: number | string
  label?: string
  legendEnabled?: boolean

  //Size of map in template
  mapFrameSize?: number[]
  mapFrameUnit?: MapFrameUnit

  overrideCommonSetting?: boolean
  enableTitle?: boolean
  enableMapPrintExtents?: boolean
  enableOutputSpatialReference?: boolean
  enableQuality?: boolean
  enableMapSize?: boolean
  enableFeatureAttribution?: boolean
  enableAuthor?: boolean
  enableCopyright?: boolean
  enableLegend?: boolean
  enableScalebarUnit?: boolean
  enableMapAttribution?: boolean
  enableCustomTextElements?: boolean
  customTextElementEnableList?: CustomTextElementEnable[]
  enableNorthArrow?: boolean

  layoutTypes?: LayoutTypes
  reportTypes?: ReportTypes
  customLayoutItem?: ActiveItem
  customReportItem?: ActiveItem

  hasAuthorText?: boolean
  hasCopyrightText?: boolean
  hasLegend?: boolean
  hasTitleText?: boolean
  selectedFormatList?: FormatType[]

  wkidLabel?: string
}

export type IMPrintTemplateProperties = ImmutableObject<PrintTemplateProperties>
export interface config {
  modeType: ModeType
  printServiceType: PrintServiceType
  printTemplateType: PrintTemplateType
  useUtility?: UseUtility
  printCustomTemplate?: PrintTemplateProperties[]
  printOrgTemplate?: PrintTemplateProperties[]
  commonSetting?: PrintTemplateProperties
  formatList?: FormatType[]
  hasInitBorder?: boolean
  previewBackgroundColor: string
  enablePreview: boolean
  previewOutLine: PreviewOutLine

  supportCustomLayout?: boolean
  supportReport?: boolean
  supportCustomReport?: boolean

  defaultFormat?: FormatType
  defaultReportTemplate?: string
  defaultCustomReportItem?: ActiveItem
  defaultCustomLayoutItem?: ActiveItem
  defaultLayout?: string

  reportTemplateChoiceList?: any[]
  layoutChoiceList?: LayoutInfo[]
}

export type IMConfig = ImmutableObject<config>